Retire the legacy webdavs-1.x service now that we only support python3 and wsgidavs in version 3.x or later.
